requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
joserfc==0.12.0
python-dotenv==1.0.1

# Snowflake
snowflake-connector-python==3.9.1
snowflake-sqlalchemy==1.5.3

# AI / TTS
openai==1.30.1          # used for OpenRouter (OpenAI-compatible)
elevenlabs==1.2.2

# Audio handling
pydub==0.25.1
aiofiles==23.2.1

# AWS / DigitalOcean Spaces (S3-compatible)
boto3==1.34.101
botocore==1.34.101

# Utils
uuid==1.30
tenacity==8.3.0
loguru==0.7.2

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
